# 영화페이지 1차 수정

+++ src/main/java/com/example/movie/dto/BoardDto.java
@@ -13,7 +13,7 @@ public class BoardDto {
     private Long boardId;
     private String title;
     private String content;
-    private double goodPoint;
+    private int goodPoint;
     private String movieNo;
     private String user;
